LayoutTests/js/statement-list-register-crash-expected.txt

Tests that code generation of statement lists properly reference counts registers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS f() is true
PASS successfullyParsed is true

TEST COMPLETE
